debundler: rip normalize+split apart, skip extract/merge re-parse
normalizeJsChunks and splitFunctionParts had been wearing each other's
trenchcoat for a while: same worker, same rebuildJsChunks helper, same
transformOneJsChunk function with an emitParts boolean and a mode switch
in the worker. They are different operations now and the file layout
reflects that.

- New common/normalize.mjs owns normalizeJsChunks end to end. Single
  serial loop, no worker pool, no shared rebuild helper. Also owns the
  per-chunk path helpers (CANONICAL_CHUNK_ENTRY_FILE,
  normalizeChunkEntryFile, normalizeAssetPath, chunkIdForJsPath) since
  none of them are split-specific.
- New common/program_analysis.mjs holds the leaf bits both passes need:
  analyzeProgramShallow + the import/export/owner-decl helpers,
  EMPTY_SPLIT_PLAN, and buildChunkManifestFromAnalysis.
- split/function_parts.mjs now contains only splitFunctionParts +
  splitOneChunk + a private worker pool. transformOneJsChunk is gone;
  the emitParts boolean is gone from splitScopeHoistedChunkAst.
- function_parts_worker.mjs renamed to split_worker.mjs; no more
  workerData.mode switch.
- load_chunks.mjs's local chunkIdForJsPath / normalizeAssetPath
  duplicates now import from common/normalize.mjs.

Two cheap perf wins on the side, prompted by per-chunk timing logs:

- extractAtomicModules t.cloneNode()s the runtime AST for the lowering
  tree instead of serializing+re-parsing it, and stores originalAst
  (not originalCode) on moduleExtractionState. statementByteCountForItem
  returns 0 when no source is supplied so plan byte counts become null
  on the fast path - acceptable tradeoff for the saving.
- mergeModules clones state.originalAst on each batch instead of
  re-parsing state.originalCode. Merge already takes a batch of
  operations as one pipeline step.

Stage timings on the benchmark: 47.9s baseline -> 11.7s now (was 15.0s
before this commit). extract_atomic_modules 12790ms -> 9926ms,
merge_modules x20 1110ms -> 687ms. Remaining bottleneck is
analyzeRuntimeBoundaryAst (7.4s on the entry chunk).

BAZEL_TEST_INVOCATIONS=buildbuddy:4caad5d0-a6f5-4805-bb39-dc38104fd610
