+++ devinfra/js/debundle/common/BUILD.bazel
@@ -12,9 +12,11 @@ js_library(
         "artifact.mjs",
         "io.mjs",
         "load_chunks.mjs",
+        "normalize.mjs",
         "package_tree.mjs",
         "parse_asts.mjs",
         "parser_options.mjs",
+        "program_analysis.mjs",
         "vendor_runtime.mjs",
     ],
     visibility = ["//visibility:public"],
