+++ devinfra/js/debundle/common/normalize.mjs
@@ -1,0 +1,148 @@
+// Pure metadata normalize: walks the loaded chunk artifact, renames every
+// chunk's entry file to a canonical "entry.js" and computes the per-chunk
+// manifest from a single shallow pass over `ast.program.body`. Reuses the
+// loaded ASTs in place — no clone, no babel-traverse, no source-code
+// re-emit. Independent of split/: nothing here cares about parts, dep
+// graphs, or scope hoisting.
+
+import { posix } from "node:path";
+import {
+  createEmptyArtifact,
+  createFile,
+  createChunk,
+  listChunks,
+  replaceChunks,
+  requirePipelineArtifact,
+  setArtifactChunkManifest,
+  setArtifactManifest,
+} from "./artifact.mjs";
+import { formatDuration, logProgress } from "./io.mjs";
+import { analyzeProgramShallow, buildChunkManifestFromAnalysis } from "./program_analysis.mjs";
+
+export const CANONICAL_CHUNK_ENTRY_FILE = "entry.js";
+
+const NORMALIZE_HEADER_LINES = Object.freeze([
+  "// Generated by //devinfra/js/debundle/split:split_chunk.",
+  "// Executable chunk entry.",
+]);
+
+export async function normalizeJsChunks({ artifact, jobs = undefined, entryFile = undefined } = {}) {
+  if (entryFile !== undefined) {
+    throw new Error("normalizeJsChunks no longer accepts entryFile; normalized chunks always use entry.js");
+  }
+  if (jobs !== undefined && jobs !== 1) {
+    logProgress(`normalize ignoring jobs=${jobs}; normalize is metadata-only and runs serially`);
+  }
+  requirePipelineArtifact(artifact, "normalizeJsChunks");
+
+  const sourceChunks = listChunks(artifact);
+  const startedAt = process.hrtime.bigint();
+  logProgress(`normalize start chunks=${sourceChunks.length}`);
+  const normalized = sourceChunks.map((chunk) => normalizeOneChunk(chunk));
+
+  const nextArtifact = createEmptyArtifact();
+  const chunkManifests = [];
+  const outputChunks = [];
+  for (const result of normalized) {
+    setArtifactChunkManifest(nextArtifact, result.chunkId, result.manifest);
+    chunkManifests.push(result.manifest);
+    outputChunks.push({
+      chunkId: result.chunkId,
+      entryFile: result.manifest.entryFile,
+      files: [
+        createFile({
+          path: result.manifest.entryFile,
+          ast: result.ast,
+          headerLines: NORMALIZE_HEADER_LINES.slice(),
+          parserOptions: result.manifest.parser,
+          metadata: {
+            chunkFile: result.manifest.entryFile,
+            chunkId: result.chunkId,
+            role: "entry",
+            sourcePath: result.inputPath,
+          },
+        }),
+      ],
+      metadata: { sourcePath: result.inputPath },
+    });
+  }
+  replaceChunks(nextArtifact, outputChunks);
+
+  const manifest = buildArtifactManifest(chunkManifests);
+  setArtifactManifest(nextArtifact, manifest);
+  const durationMs = Number(process.hrtime.bigint() - startedAt) / 1_000_000;
+  logProgress(`normalize done chunks=${normalized.length} duration=${formatDuration(durationMs)}`);
+  return { artifact: nextArtifact, manifest };
+}
+
+function normalizeOneChunk(artifactChunk) {
+  const entryArtifactFile = artifactChunk?.files?.find((file) => file.path === artifactChunk.entryFile);
+  if (!entryArtifactFile?.ast) {
+    throw new Error(
+      `normalizeJsChunks requires AST for chunk: ${artifactChunk?.chunkId ?? "<unknown>"}/${artifactChunk?.entryFile ?? "<entry>"}`
+    );
+  }
+  const inputPath = sourcePathForLoadedChunk(artifactChunk);
+  const analysis = analyzeProgramShallow(entryArtifactFile.ast);
+  const manifest = buildChunkManifestFromAnalysis(
+    artifactChunk.chunkId,
+    CANONICAL_CHUNK_ENTRY_FILE,
+    inputPath,
+    analysis,
+  );
+  return {
+    chunkId: artifactChunk.chunkId,
+    inputPath,
+    ast: entryArtifactFile.ast,
+    manifest,
+  };
+}
+
+function buildArtifactManifest(chunkManifests) {
+  return {
+    schemaVersion: 1,
+    counts: {
+      chunks: chunkManifests.length,
+      parts: 0,
+      splitFunctionDeclarations: 0,
+      keptTopLevelDeclarationOwners: chunkManifests.reduce((n, m) => n + (m.counts?.keptTopLevelDeclarationOwners ?? 0), 0),
+      topLevelSideEffects: chunkManifests.reduce((n, m) => n + (m.counts?.topLevelSideEffects ?? 0), 0),
+      exportAliases: chunkManifests.reduce((n, m) => n + (m.counts?.exportAliases ?? 0), 0),
+      unresolvedExports: chunkManifests.reduce((n, m) => n + (m.counts?.unresolvedExports ?? 0), 0),
+    },
+    chunks: chunkManifests.map((m) => ({ chunkId: m.chunkId, sourcePath: m.sourcePath })),
+  };
+}
+
+export function normalizeChunkEntryFile(entryFile) {
+  const normalized = posix.normalize(`${entryFile ?? ""}`.replaceAll("\\", "/"));
+  if (
+    normalized === "" ||
+    normalized === "." ||
+    normalized.startsWith("/") ||
+    normalized.split("/").includes("..") ||
+    !normalized.endsWith(".js")
+  ) {
+    throw new Error(`Invalid chunk entry file: ${entryFile}`);
+  }
+  return normalized;
+}
+
+export function normalizeAssetPath(path) {
+  const normalized = path.split("\\").join("/");
+  if (normalized === "" || normalized.startsWith("/") || normalized.split("/").includes("..")) {
+    throw new Error(`Invalid snapshot-relative JS path: ${path}`);
+  }
+  if (!normalized.endsWith(".js")) {
+    throw new Error(`Expected a .js path in JS list: ${path}`);
+  }
+  return posix.normalize(normalized);
+}
+
+export function chunkIdForJsPath(jsPath) {
+  return jsPath.slice(0, -".js".length);
+}
+
+function sourcePathForLoadedChunk(chunk) {
+  return normalizeAssetPath(chunk.metadata?.sourcePath ?? `${chunk.chunkId}.js`);
+}
