+++ devinfra/js/debundle/extract/atomic_modules_test.mjs
@@ -8,7 +8,7 @@ import { computeJsAsts } from "../common/parse_asts.mjs";
 import { getChunk, getChunkEntryFile } from "../common/artifact.mjs";
 import { DEFAULT_PARSER_OPTIONS } from "../common/parser_options.mjs";
 import { loadJsChunks } from "../common/load_chunks.mjs";
-import { normalizeJsChunks } from "../split/function_parts.mjs";
+import { normalizeJsChunks } from "../common/normalize.mjs";
 import { createWebFixtureRoots, runNodeScript, writeSnapshotFixture } from "../test_support/fixtures.mjs";
 import { runTransformSpecObject } from "../transforms/runner.mjs";
 import { writeJsTree } from "../transforms/write.mjs";
